+++ Permission/AllowPermission.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LoremIpsum\PermissionCheckerBundle\Permission;
+
+use LoremIpsum\PermissionCheckerBundle\AbstractPermission;
+
+class AllowPermission extends AbstractPermission
+{
+    public function isGranted()
+    {
+        return true;
+    }
+}