+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # PermissionCheckerBundle Changelog
 
+## [0.2.0] - 2018-10-05
+### Add
+- Add AllowPermission and DenyPermission
+
+### Change
+- Guardable getPermission() must return Permission
+
+### Remove
+- Remove __toString() from Guardable interface
+- Remove preCheck() from Permission interface
+
 ## [0.1.2] - 2018-10-04
 ### Change
 - Add getSecurityChecker and getTokenStorage methods to PermissionChecker